pyproject.toml
[tool.poetry]
name = "mathlab"
version = "0.1.0"
description = ""
authors = ["StanislavJPG <samper.stas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.10"
adrf = "0.1.4"
amqp = "5.2.0"
anyio = "4.4.0"
asgiref = "3.8.1"
async-property = "0.2.2"
attrs = "23.2.0"
autobahn = "23.6.2"
automat = "22.10.0"
beautifulsoup4 = "4.12.3"
billiard = "4.2.0"
celery = "5.4.0"
certifi = "^2024.6.2"
cffi = "1.16.0"
channels = "^4.2.0"
channels-redis = "4.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
constantly = "23.10.4"
contourpy = "1.2.1"
cryptography = "^44.0.1"
cycler = "0.12.1"
daphne = "^4.1.2"
django-elasticsearch-dsl = "8.0"
django-elasticsearch-dsl-drf = "0.22.5"
django-nine = "0.2.7"
django-redis = "5.4.0"
drf-yasg = "1.21.7"
elastic-transport = "8.13.1"
elasticsearch = "8.14.0"
elasticsearch-dsl = "8.14.0"
fonttools = "4.53.0"
gunicorn = "22.0.0"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
hyperlink = "21.0.0"
idna = "3.7"
incremental = "22.10.0"
inflection = "0.5.1"
jinja2 = "^3.1.4"
kiwisolver = "1.4.5"
kombu = "5.3.7"
libretranslatepy = "2.1.1"
lxml = "5.2.0"
markupsafe = "2.1.5"
matplotlib = "3.9.0"
mpld3 = "0.5.10"
mpmath = "1.3.0"
msgpack = "1.0.8"
multidict = "6.0.5"
python-dotenv = "^1.0.1"
translate = "^3.6.1"
psycopg = "^3.2.4"
sympy = "^1.13.3"
whitenoise = "^6.9.0"
pre-commit = "^4.1.0"
watchdog = "^6.0.0"
django-htmx = "^1.22.0"
django-tables2 = "^2.7.5"
django-crispy-forms = "^2.3"
django-widget-tweaks = "^1.5.0"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
django-allauth = "^65.4.1"
django-allauth-ui = "^1.5.3"
django-braces = "^1.16.0"
django-lifecycle = "^1.2.4"
uuid = "^1.30"
python-slugify = {extras = ["unicode"], version = "^8.0.4"}
django-render-block = "^0.10"
django-tinymce = "^4.1.0"
django-dynamic-filenames = "^1.4.0"
easy-thumbnails = "^2.10"
boringavatars = "^1.0.3"
ruff = "^0.9.7"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
quote-style = "single"
